processing-dicom-files-with-spark-on-cde/assets/requirements.txt
boto3
pydicom
Pillow
numpy
